Tokens functional component and assets controllers patch working
