+++ app/components/UI/Tokens/styles.ts
@@ -1,0 +1,76 @@
+import { StyleSheet } from 'react-native';
+import { fontStyles } from '../../../styles/common';
+import { Colors } from 'app/util/theme/models';
+
+const createStyles = (colors: Colors) =>
+  StyleSheet.create({
+    wrapper: {
+      backgroundColor: colors.background.default,
+      flex: 1,
+      minHeight: 500,
+    },
+    emptyView: {
+      backgroundColor: colors.background.default,
+      justifyContent: 'center',
+      alignItems: 'center',
+      marginTop: 50,
+    },
+    text: {
+      fontSize: 20,
+      color: colors.text.default,
+      ...fontStyles.normal,
+    },
+    add: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'center',
+    },
+    addText: {
+      fontSize: 14,
+      color: colors.primary.default,
+      ...fontStyles.normal,
+    },
+    tokensDetectedButton: {
+      flexDirection: 'row',
+      alignItems: 'center',
+      justifyContent: 'center',
+      marginTop: 16,
+    },
+    tokensDetectedText: {
+      fontSize: 14,
+      color: colors.primary.default,
+      ...fontStyles.normal,
+    },
+    footer: {
+      flex: 1,
+      paddingBottom: 30,
+      alignItems: 'center',
+      marginTop: 24,
+    },
+    balances: {
+      flex: 1,
+      justifyContent: 'center',
+      marginLeft: 20,
+    },
+    balanceFiat: {
+      color: colors.text.alternative,
+      ...fontStyles.normal,
+      textTransform: 'uppercase',
+    },
+    balanceFiatTokenError: {
+      textTransform: 'capitalize',
+    },
+    ethLogo: {
+      width: 24,
+      height: 24,
+      borderRadius: 12,
+      overflow: 'hidden',
+    },
+    emptyText: {
+      color: colors.text.alternative,
+      marginBottom: 8,
+      fontSize: 14,
+    },
+  });
+
+export default createStyles;
